refactored shared functions
