+++ internal/provider/datasources/data_source_api_tokens_test.go
@@ -5,6 +5,8 @@ import (
 	"os"
 	"testing"
 
+	"github.com/astronomer/terraform-provider-astro/internal/provider/common"
+
 	"github.com/astronomer/terraform-provider-astro/internal/clients/iam"
 
 	astronomerprovider "github.com/astronomer/terraform-provider-astro/internal/provider"
@@ -162,7 +164,7 @@ func checkApiTokens(tfVarName string, input checkApiTokensInput) resource.TestCh
 			if entityId != input.workspaceId {
 				return fmt.Errorf("expected 'entity_id' to be set to workspace_id")
 			}
-			if utils.ValidateRoleMatchesEntityType(role, "workspace") {
+			if common.ValidateRoleMatchesEntityType(role, "workspace") {
 				return fmt.Errorf("expected 'role' to be set as a workspace role")
 			}
 		}
@@ -183,7 +185,7 @@ func checkApiTokens(tfVarName string, input checkApiTokensInput) resource.TestCh
 			if entityId != input.organizationId {
 				return fmt.Errorf("expected 'entity_id' to be set to organization_id")
 			}
-			if utils.ValidateRoleMatchesEntityType(role, "organization") {
+			if common.ValidateRoleMatchesEntityType(role, "organization") {
 				return fmt.Errorf("expected 'role' to be set as an organization role")
 			}
 		}
